Video again, just to show and see #838
float to double conversion for the lat and lon in the gpx file

+++ src/gpx.cpp
@@ -15,21 +15,20 @@ QList<gpx_altitude_point_for_treadmill> gpx::open(const QString &gpx) {
     QDomDocument doc;
     doc.setContent(&input);
     QDomNodeList metadata = doc.elementsByTagName(QStringLiteral("metadata"));
-    if(metadata.size()) {
+    if (metadata.size()) {
         QDomNodeList list = metadata.at(0).childNodes();
-        for(int i=0; i<list.count(); i++) {
-            if(list.at(i).nodeName().toLower() == "video") {
+        for (int i = 0; i < list.count(); i++) {
+            if (list.at(i).nodeName().toLower() == "video") {
                 QString video = list.at(i).toElement().firstChild().nodeValue();
-                if(!video.isEmpty()) {
+                if (!video.isEmpty()) {
                     videoUrl = video;
                     qDebug() << "gpx::videoUrl " << videoUrl;
                     break;
                 }
-
             }
         }
     }
-    
+
     QDomNodeList points = doc.elementsByTagName(QStringLiteral("trkpt"));
     for (int i = 0; i < points.size(); i++) {
         QDomNode point = points.item(i);
@@ -41,9 +40,9 @@ QList<gpx_altitude_point_for_treadmill> gpx::open(const QString &gpx) {
         gpx_point g;
         // 2020-10-10T10:54:45
         g.time = QDateTime::fromString(time.text(), Qt::ISODate);
-        g.p.setAltitude(ele.text().toFloat());
-        g.p.setLatitude(lat.toFloat());
-        g.p.setLongitude(lon.toFloat());
+        g.p.setAltitude(ele.text().toDouble());
+        g.p.setLatitude(lat.toDouble());
+        g.p.setLongitude(lon.toDouble());
         this->points.append(g);
     }
